safari-534.53-branch/LayoutTests/platform/chromium/http/tests/security/contentSecurityPolicy/eval-blocked-expected.txt

CONSOLE MESSAGE: line 12: Uncaught Error: Code generation from strings disallowed for this context
CONSOLE MESSAGE: line 15: Uncaught Error: Code generation from strings disallowed for this context
